requirements.txt
fastapi
sqlmodel
pydantic-settings
httpx
requests
python-telegram-bot
uvicorn